move utils, rename data, probably some leftovers
