+++ .gitignore
@@ -158,7 +158,6 @@ cython_debug/
 #  and can be added to the global gitignore or merged into this file.  For a more nuclear
 #  option (not recommended) you can uncomment the following to ignore the entire idea folder.
 #.idea/
-/data/
 /logs/
 /output/
 /pretrained_weights/
@@ -166,3 +165,5 @@ cython_debug/
 /.environment/
 /configs/user.yaml
 /.environment.yaml
+/.vscode/
+/resources/
