+++ source/utils/docker_interfaces/object_detection.py
